Add record_tests.sh.

+++ Generator/run_tests.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
-rm -rf Build
-mkdir Build
-xcodebuild -project 'CuckooGenerator.xcodeproj' -scheme 'CuckooGenerator' -configuration 'Release' CONFIGURATION_BUILD_DIR=$(pwd)/Build clean build
+if [ "$1" != "--no-build" ]; then
+	rm -rf Build
+	mkdir Build
+	xcodebuild -project 'CuckooGenerator.xcodeproj' -scheme 'CuckooGenerator' -configuration 'Release' CONFIGURATION_BUILD_DIR=$(pwd)/Build clean build
+fi
 cd Tests
 cucumber
